Implement Article Modal Window
The first complete and stable version. Sure, some bug tweaks are in order, and color and scheme do nothing right now, but this is a good stable version to fall back to if need be, with most parts of the site working... more or less (home page still a mess).